fix: extracted Button2D class from elements to core
